Refactor e2e and unit tests
- Rename `roboletric` test package to e2e
- Add mocked test suite to run all e2e mocked tests
- Disable individual e2e tests on builds as we will run them from the suite
- Clean up OperationParametersTest and ClientCapabilitiesTest
