+++ msal/src/test/java/com/microsoft/identity/client/e2e/shadows/ShadowMsalUtils.java
@@ -20,27 +20,22 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-package com.microsoft.identity.client.robolectric.shadows;
+package com.microsoft.identity.client.e2e.shadows;
+
+import android.content.Context;
 
 import androidx.annotation.NonNull;
 
-import com.microsoft.identity.common.adal.internal.cache.StorageHelper;
+import com.microsoft.identity.client.internal.MsalUtils;
 
 import org.robolectric.annotation.Implements;
 
-import java.io.IOException;
-import java.security.GeneralSecurityException;
+@Implements(MsalUtils.class)
+public class ShadowMsalUtils {
 
-import javax.crypto.SecretKey;
-
-@Implements(StorageHelper.class)
-public class ShadowStorageHelper {
-
-    /**
-     * Fake saving key to key store as Android Key Store is not available in Robolectric
-     */
-    public void saveKeyStoreEncryptedKey(@NonNull SecretKey unencryptedKey) throws GeneralSecurityException, IOException {
-        return;
+    // mocking this to assume we have custom tab redirect activity during tests
+    public static boolean hasCustomTabRedirectActivity(@NonNull final Context context,
+                                                       @NonNull final String url) {
+        return true;
     }
-
 }
