+++ msal/src/test/java/com/microsoft/identity/client/e2e/tests/mocked/MultipleAccountAcquireTokenMockTest.java
@@ -20,7 +20,7 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-package com.microsoft.identity.client.robolectric.tests.mocked;
+package com.microsoft.identity.client.e2e.tests.mocked;
 
 public class MultipleAccountAcquireTokenMockTest extends AcquireTokenMockTest {
